back_poc_loki/user-service/pyproject.toml
[project]
name = "archetype"
version = "0.1.0"
description = "Archetype FAST API - Microservice based on User template"
authors = [
    {name = "Jonathan Aldana",email = "jonathan.aldana@axity.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic (==1.15.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "asyncpg (==0.30.0)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "databases (==0.9.0)",
    "fastapi (==0.115.11)",
    "freezegun (==1.5.1)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "iniconfig (==2.1.0)",
    "mako (==1.3.9)",
    "markupsafe (==3.0.2)",
    "packaging (==24.2)",
    "pluggy (==1.5.0)",
    "pydantic (==2.10.6)",
    "pydantic-settings (==2.8.1)",
    "pydantic-core (==2.27.2)",
    "pytest (==8.3.5)",
    "pytest-asyncio (==0.25.3)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "requests (==2.32.3)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.39)",
    "starlette (==0.46.1)",
    "typing-extensions (==4.12.2)",
    "urllib3 (==2.3.0)",
    "uvicorn (==0.34.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "expects (>=0.9.0,<0.10.0)",
    "structlog (>=25.2.0,<26.0.0)",
    "opentelemetry-api (>=1.32.1,<2.0.0)",
    "opentelemetry-sdk (>=1.32.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.32.1,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.53b1,<0.54)",
    "opentelemetry-instrumentation-logging (>=0.53b1,<0.54)",
    "opentelemetry-instrumentation-asgi (>=0.53b1,<0.54)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
